+++ lib/segment/src/spaces/metric_uint/avx2/avx_manhattan.rs
@@ -1,41 +1,40 @@
-use std::arch::aarch64::*;
+use std::arch::x86_64::*;
 
-#[target_feature(enable = "neon")]
+use crate::spaces::simple_avx::hsum256_ps_avx;
+
+#[target_feature(enable = "avx2")]
+#[target_feature(enable = "fma")]
 #[allow(unused)]
-pub unsafe fn neon_manhattan_similarity_bytes(v1: &[u8], v2: &[u8]) -> f32 {
+pub unsafe fn avx_manhattan_similarity_bytes(v1: &[u8], v2: &[u8]) -> f32 {
     debug_assert!(v1.len() == v2.len());
     let mut ptr1: *const u8 = v1.as_ptr();
     let mut ptr2: *const u8 = v2.as_ptr();
 
-    let mut sum16_low = vdupq_n_u16(0);
-    let mut sum16_high = vdupq_n_u16(0);
+    // sum accumulator for 8x32 bit integers
+    let mut acc = _mm256_setzero_si256();
+    // mask to take only lower 8 bits from 16 bits
+    let mask_epu16_epu8 = _mm256_set1_epi16(0xFF);
+    // mask to take only lower 16 bits from 32 bits
+    let mask_epu32_epu16 = _mm256_set1_epi32(0xFFFF);
     let len = v1.len();
-    for _ in 0..len / 16 {
-        let p1 = vld1q_u8(ptr1);
-        let p2 = vld1q_u8(ptr2);
-        ptr1 = ptr1.add(16);
-        ptr2 = ptr2.add(16);
+    for _ in 0..len / 32 {
+        // load 32 bytes
+        let p1 = _mm256_loadu_si256(ptr1 as *const __m256i);
+        let p2 = _mm256_loadu_si256(ptr2 as *const __m256i);
+        ptr1 = ptr1.add(32);
+        ptr2 = ptr2.add(32);
 
-        let abs_diff = vabdq_u8(p1, p2);
-        let abs_diff16_low = vmovl_u8(vget_low_u8(abs_diff));
-        let abs_diff16_high = vmovl_u8(vget_high_u8(abs_diff));
-
-        sum16_low = vaddq_u16(sum16_low, abs_diff16_low);
-        sum16_high = vaddq_u16(sum16_high, abs_diff16_high);
+        let sad = _mm256_sad_epu8(p1, p2);
+        acc = _mm256_add_epi32(acc, sad);
     }
-    // Horizontal sum of 16-bit integers
-    let sum32_low = vpaddlq_u16(sum16_low);
-    let sum32_high = vpaddlq_u16(sum16_high);
-    let sum32 = vaddq_u32(sum32_low, sum32_high);
 
-    let sum64_low = vadd_u32(vget_low_u32(sum32), vget_high_u32(sum32));
-    let sum64_high = vpadd_u32(sum64_low, sum64_low);
-    let mut score = vget_lane_u32(sum64_high, 0) as f32;
+    let mul_ps = _mm256_cvtepi32_ps(acc);
+    let mut score = hsum256_ps_avx(mul_ps);
 
-    let mut remainder = len % 16;
+    let mut remainder = len % 32;
     if remainder != 0 {
         let mut remainder_score = 0;
-        for _ in 0..len % 16 {
+        for _ in 0..len % 32 {
             let v1 = *ptr1 as i32;
             let v2 = *ptr2 as i32;
             ptr1 = ptr1.add(1);
@@ -50,20 +49,13 @@ pub unsafe fn neon_manhattan_similarity_bytes(v1: &[u8], v2: &[u8]) -> f32 {
 
 #[cfg(test)]
 mod tests {
-    use std::arch::is_aarch64_feature_detected;
+    use crate::spaces::metric_uint::simple_manhattan::manhattan_similarity_bytes;
 
     use super::*;
 
-    fn manhattan_similarity_bytes(v1: &[u8], v2: &[u8]) -> f32 {
-        -v1.iter()
-            .zip(v2)
-            .map(|(a, b)| (*a as i32 - *b as i32).abs())
-            .sum::<i32>() as f32
-    }
-
     #[test]
-    fn test_spaces_sse() {
-        if is_aarch64_feature_detected!("neon") {
+    fn test_spaces_avx() {
+        if is_x86_feature_detected!("avx") && is_x86_feature_detected!("fma") {
             let v1: Vec<u8> = vec![
                 255, 255, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255,
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3,
@@ -81,7 +73,7 @@ mod tests {
                 240, 239, 238,
             ];
 
-            let dot_simd = unsafe { neon_manhattan_similarity_bytes(&v1, &v2) };
+            let dot_simd = unsafe { avx_manhattan_similarity_bytes(&v1, &v2) };
             let dot = manhattan_similarity_bytes(&v1, &v2);
             assert_eq!(dot_simd, dot);
         } else {
