+++ lib/segment/src/spaces/metric_uint/avx2/mod.rs
@@ -1,0 +1,4 @@
+pub mod avx_dot;
+pub mod avx_cosine;
+pub mod avx_manhattan;
+pub mod avx_euclid;
