+++ lib/segment/src/spaces/metric_uint/simple_cosine.rs
@@ -22,7 +22,7 @@ impl Metric<VectorElementTypeByte> for CosineMetric {
     }
 }
 
-fn cosine_similarity_bytes(
+pub fn cosine_similarity_bytes(
     v1: &[VectorElementTypeByte],
     v2: &[VectorElementTypeByte],
 ) -> ScoreType {
