+++ lib/segment/src/spaces/metric_uint/simple_dot.rs
@@ -22,7 +22,7 @@ impl Metric<VectorElementTypeByte> for DotProductMetric {
     }
 }
 
-fn dot_similarity_bytes(v1: &[VectorElementTypeByte], v2: &[VectorElementTypeByte]) -> ScoreType {
+pub fn dot_similarity_bytes(v1: &[VectorElementTypeByte], v2: &[VectorElementTypeByte]) -> ScoreType {
     let mut dot_product = 0;
 
     for (a, b) in v1.iter().zip(v2) {
