+++ lib/segment/src/spaces/metric_uint/simple_manhattan.rs
@@ -22,7 +22,7 @@ impl Metric<VectorElementTypeByte> for ManhattanMetric {
     }
 }
 
-fn manhattan_similarity_bytes(
+pub fn manhattan_similarity_bytes(
     v1: &[VectorElementTypeByte],
     v2: &[VectorElementTypeByte],
 ) -> ScoreType {
