--- lib/simd/sse/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "sse"
-version = "0.0.0"
-authors = [
-    "Andrey Vasnetsov <vasnetsov93@gmail.com>",
-    "Qdrant Team <info@qdrant.tech>",
-]
-license = "Apache-2.0"
-edition = "2021"
-publish = false
-
-[dependencies]
